refactor(providers): give Cloudflare detection one owner
The first commit fixed Miruro's classifier in place, which left the repo
with two predicates answering the same question in different ways —
`isCloudflareChallengeText` in `shared/curl-impersonate.ts` (used by AniDB)
and a new marker-based one local to the Miruro module. That is how the bug
arose in the first place: Miruro grew its own copy of a concept that already
had a home, and the copy was wrong.

Both now come from one marker table in `shared/curl-impersonate.ts`, next to
the curl-candidate resolution they belong with:

- `isCloudflareChallengeText` — the interactive "Just a moment" challenge
  only. Behaviour for AniDB's four callsites is unchanged.
- `isCloudflareBlockBody` — any Cloudflare interstitial, challenge or hard
  block. Miruro uses this, because for the pipe the question is "did
  Cloudflare stop us", not "can a better fingerprint get through".

The narrow one stays narrow on purpose, and a test pins the relationship:
a challenge satisfies both, a 1020-style block satisfies only the broad one
(no retry clears it), and the origin's own `502 upstream unreachable` page
satisfies neither.

Verified: typecheck 14/14, test 23/23 with 0 fail, lint clean, and
`test:live:miruro` still resolves through `moo` on www.animegg.org.
